choose master version for following conflicts:

 Conflicts:
	InnerDetector/InDetCalibAlgs/PixelCalibAlgs/CMakeLists.txt
	InnerDetector/InDetEventCnv/PixelRawDataByteStreamCnv/CMakeLists.txt
	LArCalorimeter/LArG4/LArG4Barrel/CMakeLists.txt
	LArCalorimeter/LArG4/LArG4Barrel/LArG4Barrel/LArBarrelGeometry.h
	LArCalorimeter/LArG4/LArG4Barrel/LArG4Barrel/LArBarrelPresamplerGeometry.h
	LArCalorimeter/LArG4/LArG4Barrel/python/LArG4BarrelConfig.py
	LArCalorimeter/LArG4/LArG4Barrel/python/LArG4BarrelConfigDb.py
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelCalculator.cc
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelCalculator.h
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelCalibrationCalculator.cc
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelCalibrationCalculator.h
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelGeometry.cc
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelPresamplerCalculator.cc
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelPresamplerCalculator.h
	LArCalorimeter/LArG4/LArG4Barrel/src/LArBarrelPresamplerGeometry.cc
	LArCalorimeter/LArG4/LArG4Barrel/src/PresamplerCalibrationCalculator.cc
	LArCalorimeter/LArG4/LArG4Barrel/src/PresamplerCalibrationCalculator.h
	LArCalorimeter/LArG4/LArG4Barrel/src/components/LArG4Barrel_entries.cxx
	LArCalorimeter/LArG4/LArG4Code/LArG4Code/ILArCalculatorSvc.h
	LArCalorimeter/LArG4/LArG4Code/LArG4Code/ILArCalibCalculatorSvc.h
	LArCalorimeter/LArG4/LArG4Code/LArG4Code/LArCalculatorSvcImp.h
	LArCalorimeter/LArG4/LArG4Code/LArG4Code/LArCalibCalculatorSvcImp.h
	LArCalorimeter/LArG4/LArG4Code/src/LArCalculatorSvcImp.cxx
	LArCalorimeter/LArG4/LArG4Code/src/VolumeUtils.cc
	LArCalorimeter/LArG4/LArG4EC/CMakeLists.txt
	LArCalorimeter/LArG4/LArG4EC/cmt/requirements
	LArCalorimeter/LArG4/LArG4EC/python/LArG4ECConfig.py
	LArCalorimeter/LArG4/LArG4EC/python/LArG4ECConfigDb.py
	LArCalorimeter/LArG4/LArG4EC/src/CryostatCalibrationLArCalculator.cc
	LArCalorimeter/LArG4/LArG4EC/src/LArEndcapPresamplerCalculator.cc
	LArCalorimeter/LArG4/LArG4EC/src/LArEndcapPresamplerCalculator.h
	LArCalorimeter/LArG4/LArG4EC/src/PresamplerCalibrationCalculator.cc
	LArCalorimeter/LArG4/LArG4EC/src/PresamplerGeometry.cc
	LArCalorimeter/LArG4/LArG4EC/src/components/LArG4EC_entries.cxx
	LArCalorimeter/LArG4/LArG4GenShowerLib/LArG4GenShowerLib/TestActionShowerLib.h
	LArCalorimeter/LArG4/LArG4MiniFCAL/CMakeLists.txt
	LArCalorimeter/LArG4/LArG4MiniFCAL/src/MiniFCALCalculator.cc
	LArCalorimeter/LArG4/LArG4MiniFCAL/src/MiniFCALCalculator.h
	Projects/Athena/externals.txt
	Simulation/G4Atlas/G4AtlasAlg/src/AthenaTrackingAction.cxx
	Simulation/G4Atlas/G4AtlasApps/python/PyG4Atlas.py
	Simulation/G4Atlas/G4AtlasApps/python/SimAtlasKernel.py
	Simulation/G4Atlas/G4AtlasTests/src/MMHitsTestTool.cxx
	Simulation/G4Atlas/G4AtlasTests/src/MMHitsTestTool.h
	Simulation/G4Atlas/G4AtlasTests/src/MuonHitTestToolBase.cxx
	Simulation/G4Atlas/G4AtlasTests/src/MuonHitTestToolBase.h
	Simulation/G4Atlas/G4AtlasTests/src/sTGCHitsTestTool.cxx
	Simulation/G4Atlas/G4AtlasTests/src/sTGCHitsTestTool.h
	Simulation/G4Atlas/G4AtlasTools/src/SensitiveDetectorBase.cxx
	Simulation/G4Utilities/GeoMaterial2G4/GeoMaterial2G4/Geo2G4MatPropTableFactory.h
	Simulation/G4Utilities/GeoMaterial2G4/GeoMaterial2G4/Geo2G4MaterialFactory.h
	Simulation/G4Utilities/GeoMaterial2G4/cmt/requirements
	Simulation/G4Utilities/GeoMaterial2G4/src/Geo2G4ElementFactory.h
	Simulation/G4Utilities/GeoMaterial2G4/src/Geo2G4MatPropTableFactory.cxx
	Simulation/G4Utilities/GeoMaterial2G4/src/Geo2G4MaterialFactory.cxx


Former-commit-id: 74b9b40485578f2a7690f9b34d04529351b328b0

+++ Trigger/TrigAnalysis/TrigInDetAnalysisUser/scripts/TIDArun.sh
@@ -35,53 +35,58 @@ if [ -e TrkNtuple-0000.root ]; then
     fi
 
     echo "fetching reference files from $REFDIR"
-     
-
-    EXPERT=$(grep expert $2) 
+        
+    EXPERT=$(echo $1 $2 | grep expert) 
 
     # if expert timing histos
     if [ "x$EXPERT" != "x" ]; then 
-      REFFILE=$REFDIR/expert/$2
+      if [ "x$RTTJOBNAME" != "x" ]; then 
+        EXPERTREF="expert-monitoring-$(echo $RTTJOBNAME | sed 's|TrigInDetValidation_||g')-ref.root"
+        REFFILE=$REFDIR/expert/$EXPERTREF
+      else
+        REFFILE=$REFDIR/expert/$2
+      fi      
     else
       REFFILE=$REFDIR/$2
     fi
 
-    if [ -e  $REFFILE ]; then 
+    NOREF=
+
+    if [ ! -e $REFFILE ]; then 
+
+      echo "no reference file $REFFILE - will process without"
+
+      NOREF=--noref
+
+    else
+
       echo "copying $REFFILE" 
 
       LOCALFILE=$(basename $REFFILE)
 
-      # don;t bother to copy it if it already exists in this directory
-      if [ ! -e $LOCALFILE ]; then 
-         cp $REFFILE .
-      fi
+      # don't bother to copy it if it already exists in this directory
+      [ -e $LOCALFILE ] || cp $REFFILE .
+  
+    fi
 
-      ls -l
+    ls -l
 
-      if [ "x$EXPERT" == "x" ]; then 
+    for arg in $*; do 
+       if $(echo "$arg" | grep -q "TIDA.*.dat"); then get_files -data "$arg"; fi 
+       if $(echo "$arg" | grep -q "Test.*.dat"); then get_files -data "$arg"; fi 
+    done
 
-        echo
-        echo "running comparitor " `date`
-        echo
-
-        TIDAcomparitor.exe $*
-
-      else
-
-        echo
-        echo "running cpucost " `date`
-        echo
-
-        TIDAcpucost.exe $*
-
-      fi
-
+    if [ "x$EXPERT" == "x" ]; then 
+        get_files -data TIDAhistos-vtx.dat
+        echo -e "\nrunning comparitor " $(date) "\n"
+        TIDAcomparitor.exe $* $NOREF
     else
-      echo "no reference file $REFFILE" 
+        echo -e "\nrunning cpucost " $(date) "\n"
+        TIDAcpucost.exe $* $NOREF
     fi
 
     echo "finished postprocessing"
-    ls -l
 
+    ls -l
 
 fi
